Added item filters in MassQueryConfig
