--- app/include/database/MassQueryResult.hpp
@@ -1,21 +1,0 @@
-#pragma once
-
-#include <memory>
-#include <vector>
-
-#include "electrical/ElectronicComponent.hpp"
-
-namespace ecim {
-    /// @brief The MassQueryResult struct represents a results of a mass query
-    /// executed on the database.
-    ///
-    /// Each instance contains a list of electronic components matching the
-    /// criteria specified by the query, as well as metadata such as the total
-    /// number of items returned, and the total number of pages that could be
-    /// obtained if pagination is enabled.
-    struct MassQueryResult {
-        std::vector<std::unique_ptr<ElectronicComponent>> items;
-        size_t totalItems;
-        size_t totalPages;
-    };
-}
